Use on-chain registry for banned users (#2443)
# Description
This PR adds support to enforce an onchain registry of banned users.
Previously those could only be supplied on startup via a hardcoded list.
The list behavior is retained and extended by an optional on-chain
lookup. The result of the on-chain lookup is cached for 1h.

# Changes
- [x] Create new shared crate that implements banned user logic
- [x] Use [ttl-cache](https://crates.io/crates/ttl_cache) crate for
cache implemation
- [x] Replace basic `Hashset::contains` check if `autopilot` and
`orderbook` with new async check of the new component

50% of the lines are the copied licence files.

## How to test
Existing unit tests. Will write a forked e2e test to test the on-chain
behavior.