Send extra fields from driver to solving engine (#2488)
# Description
Solving engines need to be aware of how much gas their solutions use to
set the feeAmounts correctly. To do so properly, they need to be able to
simulate transactions. To do so, they need to know some extra fields
that were not passed on by the driver:
- router for ConstantProductPool and ConcentratedLiquidityPool
- balancer_pool_id for StablePool and WeightedProductPool

# Changes
- added missing fields in structs in auction
- set the fields when creating Auction struct 

## How to test
trivial changes, not sure if or where to add tests

Co-authored-by: Felix Leupold <felixleupold90@gmail.com>