Optional gas estimate on solution (#2524)
# Description
Fixes: https://github.com/cowprotocol/services/issues/2516

# Changes
Introduce `gas` field to let solvers tell us how much gas they estimate
the execution of the quote or settlement to cost. Now if a solver does
not provide a gas estimate but we can also not estimate the gas on our
own by simulating we discard the quote entirely.

Note that I didn't add any gas estimation logic for the `naive` solver
since we currently only need the `gas` estimate for quotes which `naive`
is not able to generate.
For good support we also need to patch the `solvers` repository to add
gas estimates provided by dex APIs in their solution.

## How to test
Adjusted e2e tests to make them pass again.