Add log for fetched events (#2518)
# Description

Related to oncall issue where our system failed to index some
settlements/trades. The only reason I could think of at the moment is
that it was some intermittent issue with the node which answered with
empty Vec of events. All other error execution paths are already covered
in the code.