Baseline solution offset (#2565)
# Description
Since we moved to rank by surplus baseline is bleeding money due to bad
gas estimates for its solutions which leads to collecting not enough
fees.
These missing fees eventually have to be paid out of the solver's
pocket.

# Changes
Add a config parameter that allows increasing or decreasing a solution's
gas estimate. According to @harisang this should be enough to stop the
bleeding.
Obviously this is very hacky and we should a better solution soon which
very likely involves simulating the solution.