Actually sort descending 🙈 (#2623)
# Description
It turns out we were not actually sorting solutions in descending but in
ascending score when merging them. This is not an issue per se as -
given enough time - all solutions get simulated and the one with the
highest score gets picked (for this there are tests).

However, if there are a lot of solution candidates (more than we have
time to encode), the ordering matters for prioritisation.

# Changes
- [x] Reverse sorting 

## How to test
Since this aspect only happens when we run out of time it's quite hard
to test. Not sure if it's necessary. If you have an idea please let me
know.