Add request id header for rpc requests (#2625)
# Description
Add a request id header for the RPC requests.

# Changes
<!-- List of detailed changes (how the change is accomplished) -->
This PR adds the following request HTTP headers on RPC requests:
- `X-RPC-Request-Id` - The unique ID of the RPC request.
- `X-Request-Id` - The top level request id passed as a request header
for the RPC request generated.

## How to test
<!--- Include details of how to test your changes, including any
pre-requisites. If no unit tests are included, please explain why and
how to test manually
1.
2.
3.
-->
I couldn't find a suitable way to test this. Maybe, it would be useful
to create tests for the RPC request execution and assert that those
headers are been sending.

## Related Issues

Fixes #2519

Co-authored-by: Martin Beckmann <martin.beckmann@protonmail.com>