Move fee policy database structs (#2657)
# Description
Moves `struct FeePolicy`, used to read/write fee policies from database,
to database crate, for reuse. A prerequisite for
https://github.com/cowprotocol/services/issues/2642#issuecomment-2074781498,
as we need to expose fee policies as a new field in `trades` query, and
for this we need to reuse `FeePolicy` struct.

The code change is mostly just a move, no new functionality.

## How to test
Existing tests.