Extract custom clearing price logic (#2659)
# Description
For the domain based settlement encoding we need to encode both the
uniform clearing prices of the solution as well as the custom (fee
adjusted) clearing prices of each trade. This PR create a utility
function on `Trade` that allows this logic to be shared.

After rebasing, I noticed @Mateo-mro already introduced a similar
function (`calculate_custom_price`). However, this one wasn't reusing
the existing math to calculate net buy and sell amounts and had a
slightly longer name, so this is now mainly a refactoring/renaming of
this method.

# Changes
- [x] Move `CustomClearingPrice` type to up since it won't be `scoring`
specific
- [x] Reuse buy/sell amount calculation to derive the clearing price
- [x] Rename use of `calculate_custom_price`

## How to test
CI

## Related Issues

Part of #2215