Don't use settlement contract as token owner for bad token test (#2681)
# Description
The bad token test finds an address that holds a good amount of tokens
and tries to send some amount of that to the settlement contract and
back again. Afterwards it checks that when we send X tokens the receiver
actually got X tokens and not some smaller amount to filter out tokens
with fee on transfer behavior.

Because the token balance assertion is `if owner sends X tokens to the
settlement contract the settlement_contract has X tokens more` it
obviously fails when `owner == settlement_contract` because the balance
for good tokens will also not change flagging them incorrectly as fee on
transfer tokens.

# Changes
Filter out the settlement contract in the `TokenFinder` implementation.