Switch to simple block stream strategy (#2674)
# Description
Switch to fetching the current block via `get-block` again. Previously
we used a strategy called `get-block-and-call` which does an
`eth_getBlock` and an `eth_call` that fetches the current block via a
helper contract as a smart contract would see it and returns a
reasonable choice when these values differ.
We made this the default because it worked and we didn't run any
different strategy anymore.
This does not work on arbitrum because `eth_getBlock` and `eth_call`
return wildly different results. The reason is that `eth_getBlock`
returns the current block of the L2 chain whereas `eth_call` returns the
current block of L1. These 2 blocks have block numbers millions apart so
the code panics when it tries to reconcile the different values.
Luckily `get-block-and-call` was only introduced to work around a bug in
nethermind which would return the new block when the block headers have
been processed but not the content of the block meaning that we could
get "no state for block" errors when we tried to fetch state for the
latest block we just queried.
Since we no longer run `nethermind` nodes switching back to
`eth_getBlock` should be fine.

# Changes
Switch to `eth_getBlock` to fetch the latest block by default.

## How to test
Already ran arbitrum successfully with the `get-block` strategy.