Set gas price for quote simulations (#2697)
# Description
Fixes https://github.com/cowprotocol/services/issues/2696
Usually if you don't set a gas price for an `eth_call` "a sane default"
should be picked by the node. This either does not work with the nodes
we use or simulating with state overrides is an exception to the rule.

# Changes
Read gas price in `CurrentBlockStream` and use the gas price of the most
recent block for quote simulations.

## How to test
e2e tests should still work as the gas price should not affect the gas
used computed during the verification simulation.