[Trivial] Extend driver shutdown time (#2726)
# Description
We have been seeing panics like
[this](https://cowservices.slack.com/archives/C037PB929ME/p1715586375609859)
sporadically on system restarts.

I believe the root-cause for this is that the time we leave the driver
to gracefully shut down (10s) is shorter than the time a single solve
request may be taking. This when a solve request is in flight, we don't
wait long enough before forcing shutdown. Timeline from the example ☝️

- 2024-05-13T07:45:28.781Z driver competition started
- 2024-05-13T07:45:30.313Z likely receipt of shutdown signal (no logs)
- 2024-05-13T07:45:40.313Z panic due to timeout
- 2024-05-13T07:45:43.254Z driver competition deadline would have been
reached

# Changes
- [x] Wait 20s for timeout
- [x] Log when shutdown signal is received

## How to test
CI and hope to no longer see those panics.