Create auction orders table (#2748)
# Description
Fixes https://github.com/cowprotocol/services/issues/2745

This table is supposed to be populated for each `Auction`. All orders
that are part of the `Auction` will be saved to this table, during
competition.

This information is important so we can reconstruct the whole
competition environment needed for circuit breaker and
on_settlement_event actions.

## How to test
Added unit test. CI.