[TRIVIAL] Simplify estimator logs (#2772)
# Changes
When quote verification was fresh it made sense to have special logging
for verified estimators to debug things. But now all estimators are
always verified and the `_verified` postfix makes it so that you don't
get to see all the logs in kibana when you just search for `<estimator>`
(`<estimator>_verified` logs would not show up).