(Proposal):  Remove additional_deadline_for_rewards (#2789)
# Description
It seems the field `additional_deadline_for_rewards` is not used
anywhere. It is used to calculate the `block_deadline` which is stored
in the database as part of the competition. But it seems it has no other
effect. Maybe is it legacy code?

# Changes
Fully remove `additional_deadline_for_rewards`.

## How to test
1. Regression tests