Cow amm driver support (#2790)
# Description
Adds full support for cow amms in the backend. This includes:
* new contracts in `build.rs`
* `cow-amm` crate which encapsulates indexing of pools and generating
template orders
* autopilot adds cow amms to surplus capturing jit order owners and
indexes their native prices
* driver generates template orders for all relevant amms
* e2e test based on a fork of mainnet

Kudos to @Mateo-mro for implementing the indexing logic and coming up
with the architecture that allowed us to create a single `Registry` that
is able to index multiple different contracts to have a simpler API.
Due to some misconceptions on my part the helper contract needed to be
adjusted to make indexing of v1 cow amms possible this led to
significant changes in the `cow-amm` crate so I offer this PR to give
the whole picture instead of refactoring on top of existing PRs.

## How to test
e2e test covers that the driver is able to suggest template orders for
imbalanced cow amms and that the driver is able to settle them. Also
verifies that the autopilot adds all amms to the surplus capturing jit
order owners in the auction sent to the driver.

---------

Co-authored-by: Mateo <mateo@cow.fi>
Co-authored-by: Mateo-mro <160488334+Mateo-mro@users.noreply.github.com>