Pass CoW AMM configs via args (#2803)
# Description
Currently we try to automatically load the `CowAmmLegacyHelper` contract
in the autopilot and driver. This is convenient in terms of writing code
but does not allow to disable indexing cow amms.
This is particularly problematic if the staging and prod driver generate
orders for the same CoW AMMs. In that case these 2 environments are
racing each other which could lead to many reverts when both
environments try to rebalance the same CoW AMM.

# Changes
Now we can pass the cow amm configs (start_block, factory, helper) via
configuration parameters.
This allows us to also run the driver without generating orders in
staging and also adds the necessary features to support the balancer
pools out of the gate.
Now as soon as the helper contract is ready we can just update the
configs in the infra and start rebalancing the new pools.

## How to test
New config arguments are tested in existing e2e test