Implement CoinGecko for fetching native prices (#2824)
# Description
Implement a native prices fetch from CoinGecko. As we see in the issue
https://github.com/cowprotocol/services/issues/2814, we are getting many
orders rejected because we do not have native prices for many tokens.
And the cache of outdated tokens keeps growing over time. In order to
solve this, we add CoinGecko as a native price fetcher too.

This PR standalone (without infrastructure PR) doesn't have any impact,
as CoinGecko must be configured as a native price estimator source.

The PR can already prove if fetching prices from CoinGecko improves the
cache and the order errors rate, but the following changes are needed in
upcoming PRs:
- Instead of fetching prices from all the sources at one, have a primary
method for fetching prices (CoinGecko) and fallback methods (solvers +
1inch)
- Potentially improve performance by doing bulk requests
- Move the code from autopilot

Please, see my comments below for more clarification.

# Changes
- Add configuration for CoinGeck estimator source
- Implement the trait for fetching native prices for CoinGecko

## How to test
1. Unit test

## Related Issues

### Partially fixes: https://github.com/cowprotocol/services/issues/2814

---------

Co-authored-by: ilya <ilya@cow.fi>