Fix already recorded data for auction_id (#2842)
# Description
Second and third match cases substituted place (code diff shows it a bit
unfortunately).

`settlement data already recorded for this auction` should have a higher
priority. The example when this is important is when staging (which is
frontrunning production with `auction_id` by ~270 days) reports
`solution submitted by solver other than the winner` in cases when it
processes production tx with an `auction_id` that is already processed
by staging 270 days ago:

https://production-6de61f.kb.eu-central-1.aws.cloud.es.io/app/r/s/nsx2A

Instead, it should first check if the settlement is already processed
and skip the settlement, while `solution submitted by solver other than
the winner` is a serious violation that should indicate something is
wrong (although we currently don't act on it).